CRUD of users and books finished

+++ src/useCases/books/deleteBook.ts
@@ -1,14 +1,26 @@
-// import BooksRepository from "../../repositories/booksRepository";
+import { Book } from "@prisma/client";
+import searchErrors from "../../dtos/searchError";
+import searchExceptions from "../../exceptions/searchExceptions";
 import { prisma } from "../../prisma/Client";
 
 export default class DeleteBookUseCase {
-  // private _repository: BooksRepository
-  // constructor(repository: BooksRepository){
-  //     this._repository = repository
-  // }
+  public async execute(id: string): Promise<void> {
+    let errors: searchErrors;
+    const searchId: Book | null = await prisma.book.findFirst({
+      where: {
+        id,
+      },
+    });
 
-  public execute(id: string): void {
-    const deletedBook = prisma.book.delete({
+    if (!searchId) {
+      throw new searchExceptions(
+        (errors = {
+          message: "No books were found",
+          statusCode: 404,
+        })
+      );
+    }
+    const deletedBook = await prisma.book.delete({
       where: {
         id,
       },
